+++ crates/remote/src/grpc_remote_client.rs
@@ -300,26 +300,11 @@ impl RemoteClient for GrpcRemoteClient {
                 Ok(Some(response.into_inner()))
             }
             Err(status) => {
-                let code = status.code();
-
-                if matches!(code, Code::InvalidArgument | Code::FailedPrecondition) {
-                    warn!(
-                        hash = &action_digest.hash,
-                        code = ?code,
-                        "Failed to cache action result: {}",
-                        color::muted_light(status.message()),
-                    );
-
-                    Ok(None)
-                } else if matches!(code, Code::ResourceExhausted) {
-                    warn!(
-                        hash = &action_digest.hash,
-                        code = ?code,
-                        "Remote service is out of storage space: {}",
-                        color::muted_light(status.message()),
-                    );
-
-                    Ok(None)
+                if matches!(status.code(), Code::ResourceExhausted) {
+                    Err(RemoteError::GrpcOutOfStorageSpace {
+                        error: Box::new(status),
+                    }
+                    .into())
                 } else {
                     Err(self.map_status_error("update_action_result", status).into())
                 }
@@ -367,18 +352,7 @@ impl RemoteClient for GrpcRemoteClient {
             .await
         {
             Ok(response) => response,
-            Err(status) => {
-                return if matches!(status.code(), Code::InvalidArgument) {
-                    warn!(
-                        hash = &action_digest.hash,
-                        "Attempted to download more blobs than the allowed limit"
-                    );
-
-                    Ok(vec![])
-                } else {
-                    Err(self.map_status_error("batch_read_blobs", status).into())
-                };
-            }
+            Err(status) => return Err(self.map_status_error("batch_read_blobs", status).into()),
         };
 
         let mut blobs = vec![];
@@ -469,14 +443,10 @@ impl RemoteClient for GrpcRemoteClient {
                     bytes.extend(data.data);
                 }
                 Err(error) => {
-                    warn!(
-                        hash = &action_digest.hash,
-                        blob_hash = &blob_digest.hash,
-                        "Failed to stream download blob: {}",
-                        color::muted_light(error.to_string()),
-                    );
-
-                    return Ok(None);
+                    return Err(RemoteError::GrpcStreamDownloadFailed {
+                        error: Box::new(error),
+                    }
+                    .into());
                 }
             }
         }
@@ -536,24 +506,11 @@ impl RemoteClient for GrpcRemoteClient {
         {
             Ok(response) => response,
             Err(status) => {
-                let code = status.code();
-
-                return if matches!(code, Code::InvalidArgument) {
-                    warn!(
-                        hash = &action_digest.hash,
-                        "Attempted to upload more blobs than the allowed limit"
-                    );
-
-                    Ok(vec![])
-                } else if matches!(code, Code::ResourceExhausted) {
-                    warn!(
-                        hash = &action_digest.hash,
-                        code = ?code,
-                        "Remote service exhausted resource: {}",
-                        color::muted_light(status.message()),
-                    );
-
-                    Ok(vec![])
+                return if matches!(status.code(), Code::ResourceExhausted) {
+                    Err(RemoteError::GrpcOutOfStorageSpace {
+                        error: Box::new(status),
+                    }
+                    .into())
                 } else {
                     Err(self.map_status_error("batch_update_blobs", status).into())
                 };
@@ -648,15 +605,11 @@ impl RemoteClient for GrpcRemoteClient {
 
         let result = self.get_bs_client().write(Request::new(stream)).await;
 
-        if let Some(ref error) = *stream_error.lock().await {
-            warn!(
-                hash = &action_digest.hash,
-                blob_hash = &blob.digest.hash,
-                "Failed to stream upload blob: {}",
-                color::muted_light(error.to_string()),
-            );
-
-            return Ok(None);
+        if let Some(error) = Arc::into_inner(stream_error).and_then(|error| error.into_inner()) {
+            return Err(RemoteError::GrpcStreamUploadFailed {
+                error: Box::new(error),
+            }
+            .into());
         }
 
         match result {